FIX strong validation for render state - batch; flush needed before state changes - fbo; i.e. water map rendering problems

+++ src/asset/Assets.h
@@ -41,7 +41,7 @@ public:
     float gbufferScale;
     float bufferScale;
 
-    glm::vec2 windowSize;
+    glm::uvec2 windowSize;
     bool windowMaximized;
 
     glm::vec3 cameraMoveNormal;
